+++ src/common/axios.js
@@ -1,0 +1,9 @@
+import axios from "axios";
+
+export default axios.create({
+   baseURL: "http://localhost:8080",
+   headers: {
+      "Content-Type": "application/json"
+   },
+   withCredentials: true,
+});
